Le type inet_addr est represente comme une chaine, pour pouvoir passer dans output_value.
git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@383 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ otherlibs/unix/socketaddr.h
@@ -29,4 +29,4 @@ void get_sockaddr P((value));
 value alloc_sockaddr P((void));
 value alloc_inet_addr P((unsigned int));
 
-#define GET_INET_ADDR(v) (*((unsigned long *) (v)))
+#define GET_INET_ADDR(v) (*((uint32 *) (v)))
